+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -23,7 +23,7 @@
 ## Version 1.3.2
 * Introduced Get-AzCosmosDBMongoDBBackupInformation to retrieve latest backup information for MongoDB.
 * Update New-AzCosmosDBAccount,Update-AzCosmosDBAccount to accept BackupStorageRedundancy
-* Introduced Get-AzCosmosDBLocation to list supported CosmosDB locations and properties.
+* Introduced Get-AzCosmosDBLocation to list Azure CosmosDB Account and its locations properties.
 
 ## Version 1.3.1
 * Fixed a bug where the restore of deleted database accounts fail.